requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
click==8.2.1
fastapi==0.115.12
flake8==7.2.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.13.0
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
